enjoy

+++ Includes/cub3d.h
@@ -157,7 +157,7 @@ int		check_init_done(t_data *d);
 void	print_all(t_data *data);
 void	free_cube(t_data *d);
 void	message(char *msg, int n, t_data *d);
-bool	is_decimal(float n);
+bool	is_decimal(double n);
 int		close_window(t_data *d);
 
 // draw
@@ -165,7 +165,7 @@ int		draw_map(t_data *d);
 void	draw_minimap(t_data *d);
 void	draw_minimap(t_data *d);
 void	draw_dashed_line(t_data *d, int p_pos_x, int p_pos_y);
-void	draw_ray(t_data *d, float x_center, float y_center);
+void	draw_ray(t_data *d, double x_center, double y_center);
 void	draw_player(t_data *d, double pos_x, double pos_y);
 
 // move
